+++ contracts/voting/dao-voting-cw721-staked/src/testing/mod.rs
@@ -3,6 +3,7 @@ mod execute;
 mod hooks;
 mod instantiate;
 mod queries;
+mod stargaze;
 mod tests;
 
 use cosmwasm_std::Addr;
