+++ packages/dao-interface/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod migrate_msg;
 pub mod msg;
+pub mod nft;
 pub mod proposal;
 pub mod query;
 pub mod state;
