+++ packages/dao-interface/src/nft.rs
@@ -1,0 +1,6 @@
+use cosmwasm_schema::cw_serde;
+
+#[cw_serde]
+pub struct NftFactoryCallback {
+    pub nft_contract: String,
+}
